Combined to single .h file
I ran into some issues getting the ifdefs to exclude unwanted code - I want to
minimize the footprint for the AVR processors as the library starts to support more
platforms.
It turns out that the only way to get this to work in the normal user Arduino IDE
environment - where people don't tweak their own compiler options - is to slam it all
into the .h file.

Trivially tested.